// types.ts
export type Role =
  | "SuperAdmin"
  | "Admin"
  | "Doctor"
  | "Nurse"
  | "Receptionist"
  | "Lab Tech"
  | "Pharmacist"
  | "Accountant"

export interface TeamMember {
  id: string
  clinicId?: string
  name: string
  email: string
  phone?: string
  role: Role
  status: "Active" | "Invited" | "Deactivated"
  lastActive: string
  avatar?: string
}